requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
importlib-metadata==3.7.3
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
python-dotenv==0.15.0
selenium==3.141.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1